+++ src/main/java/lotto/domain/lottoresult/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottoresult;
 
 public interface LottoResult {
