+++ src/main/java/lotto/rankingpolicy/RankingPolicy.java
@@ -1,8 +1,7 @@
 package lotto.rankingpolicy;
 
 import lotto.domain.LottoRank;
-import lotto.domain.DefaultLottoResult;
-import lotto.domain.LottoResult;
+import lotto.domain.lottoresult.LottoResult;
 
 public interface RankingPolicy {
     LottoRank grade(LottoResult lottoResult);
