+++ src/test/java/lotto/domain/DefaultLottoResultTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.lottoresult.DefaultLottoResult;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
